+++ src/pages/Unlock/hooks/index.ts
@@ -1,3 +1,4 @@
+export * from './useGetNativeAuthConfig';
 export * from './useInitToken';
 export * from './useOnFileLogin';
 export * from './useOnLoginHookRedirect';