+++ src/pages/index.tsx
@@ -3,6 +3,7 @@ export * from './SovereignTransfer';
 export * from './Disclaimer';
 export * from './Home';
 export * from './Hook';
+export * from './IssueToken';
 export * from './Logout';
 export * from './PageNotFound';
 export * from './Send';